Polish sidebar atomic selectors and controllers
