+++ apps/web/src/components/Sidebar.logic.ts
@@ -20,6 +20,12 @@ type SidebarProject = {
   updatedAt?: string | undefined;
 };
 
+type SidebarThreadSortInput = Pick<Thread, "createdAt" | "updatedAt"> & {
+  projectId: string;
+  latestUserMessageAt?: string | null;
+  messages?: Pick<Thread["messages"][number], "createdAt" | "role">[];
+};
+
 export type ThreadTraversalDirection = "previous" | "next";
 
 export interface ThreadStatusPill {
@@ -46,15 +52,27 @@ const THREAD_STATUS_PRIORITY: Record<ThreadStatusPill["label"], number> = {
 
 type ThreadStatusInput = Pick<
   SidebarThreadSummary,
-  | "hasActionableProposedPlan"
-  | "hasPendingApprovals"
-  | "hasPendingUserInput"
-  | "interactionMode"
-  | "latestTurn"
-  | "session"
+  "hasActionableProposedPlan" | "hasPendingApprovals" | "hasPendingUserInput" | "interactionMode"
 > & {
+  latestTurn: ThreadStatusLatestTurnSnapshot | null;
+  session: ThreadStatusSessionSnapshot | null;
   lastVisitedAt?: string | undefined;
 };
+
+type LatestTurnSnapshot = NonNullable<SidebarThreadSummary["latestTurn"]>;
+type SessionSnapshot = NonNullable<SidebarThreadSummary["session"]>;
+
+export interface ThreadStatusLatestTurnSnapshot {
+  turnId: LatestTurnSnapshot["turnId"];
+  startedAt: LatestTurnSnapshot["startedAt"];
+  completedAt: LatestTurnSnapshot["completedAt"];
+}
+
+export interface ThreadStatusSessionSnapshot {
+  orchestrationStatus: SessionSnapshot["orchestrationStatus"];
+  activeTurnId?: SessionSnapshot["activeTurnId"];
+  status: SessionSnapshot["status"];
+}
 
 export interface ThreadJumpHintVisibilityController {
   sync: (shouldShow: boolean) => void;
@@ -489,7 +507,7 @@ export function getProjectSortTimestamp(
 
 export function sortProjectsForSidebar<
   TProject extends SidebarProject,
-  TThread extends Pick<Thread, "projectId" | "createdAt" | "updatedAt"> & ThreadSortInput,
+  TThread extends SidebarThreadSortInput,
 >(
   projects: readonly TProject[],
   threads: readonly TThread[],
