+++ apps/web/src/components/sidebar/sidebarConstants.ts
@@ -1,0 +1,1 @@
+export const THREAD_PREVIEW_LIMIT = 6;
