+++ apps/web/src/logicalProject.ts
@@ -2,9 +2,11 @@ import { scopedProjectKey, scopeProjectRef } from "@t3tools/client-runtime";
 import type { ScopedProjectRef } from "@t3tools/contracts";
 import type { Project } from "./types";
 
+export type LogicalProjectKey = string;
+
 export function deriveLogicalProjectKey(
   project: Pick<Project, "environmentId" | "id" | "repositoryIdentity">,
-): string {
+): LogicalProjectKey {
   return (
     project.repositoryIdentity?.canonicalKey ??
     scopedProjectKey(scopeProjectRef(project.environmentId, project.id))
@@ -14,6 +16,6 @@ export function deriveLogicalProjectKey(
 export function deriveLogicalProjectKeyFromRef(
   projectRef: ScopedProjectRef,
   project: Pick<Project, "repositoryIdentity"> | null | undefined,
-): string {
+): LogicalProjectKey {
   return project?.repositoryIdentity?.canonicalKey ?? scopedProjectKey(projectRef);
 }
